% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffsig-package.R
\docType{package}
\name{diffsig-package}
\alias{diffsig-package}
\alias{diffsig}
\title{The 'diffsig' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2021). RStan: the R interface to Stan. R package version 2.21.3. https://mc-stan.org
}
