% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffsig_eval.R
\name{diffsig_eval}
\alias{diffsig_eval}
\title{Diffsig coverage and root mean squared error (RMSE) calculation function}
\usage{
diffsig_eval(truebeta, fit, ci_level = 80, include = T, ...)
}
\arguments{
\item{truebeta}{a vector of true betas}

\item{fit}{a \code{stanfit} object returned by \code{rstan::stan}}

\item{...}{Arguments passed to \code{rstan::stan} (e.g. thin, init, ...).}

\item{coverage}{a 2-dimension vector of lower/upper credible interval}

\item{beta_sd}{standard deviation for sampling hyperparameter beta}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::stan}
}
\description{
Diffsig coverage and root mean squared error (RMSE) calculation function
}
