% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffsig_fit.R
\name{diffsig_fit}
\alias{diffsig_fit}
\title{Hierarchical Bayesian Modeling with Stan}
\usage{
diffsig_fit(X, Y, C, beta_sd, pars = c("beta"), include = T, ...)
}
\arguments{
\item{C}{muational signature matrix (LxK)}

\item{beta_sd}{standard deviation for sampling hyperparameter beta}

\item{pars}{desired parameters from \code{stanfit} to save. Default is "beta" (pars="beta"), but can also include parameters such as "tau".}

\item{...}{Arguments passed to \code{rstan::stan} (e.g. thin, init, ...).}

\item{x}{risk factor matrix (MxN)}

\item{y}{mutational count matrix (LxN)}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::stan}
}
\description{
Hierarchical Bayesian Modeling with Stan
}
