% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_trans.R
\name{select_trans}
\alias{select_trans}
\title{Selecting L mutation transitions out of 96}
\usage{
select_trans(Y, C, L)
}
\arguments{
\item{Y}{mutation count matrix Y (96xN)}

\item{C}{mutation signature matrix C (96xK)}

\item{L}{number of mutation transitions to be reduced to}
}
\value{
list of subsetted signature matrix C and subsetted count data Y (=list(C,Y))
}
\description{
Selecting L mutation transitions out of 96
}
