% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdat_gen.R
\name{simdat_gen}
\alias{simdat_gen}
\title{Simulation data generator to test diffsig}
\usage{
simdat_gen(
  N = 100,
  K = 5,
  L = 96,
  truetau = 100,
  C,
  continuous = NA,
  binary = NA,
  categorical = NA,
  cat_levels = NA,
  seed = sample(1:1e+06, 1)
)
}
\arguments{
\item{N}{number of samples, default 100}

\item{K}{number of signatures, default 5}

\item{L}{number of transitions to use, default 96}

\item{truetau}{true value of Dirichlet precision parameter tau, default 100}

\item{C}{mutational signature matrix, COSMIC signatures recommended (LxK)}

\item{continuous}{number of continuous variables}

\item{binary}{number of binary variables}

\item{categorical}{number of categorical variables}

\item{cat_levels}{vector of number of levels for each categorical variables}

\item{seed}{seed number for random number generate (set.seed)}

\item{M}{number of covariates, default 2 (includes intercept)}
}
\value{
A list of simulated beta, mutational count (Y), and risk factor (X)
}
\description{
Simulation data generator to test diffsig
}
