% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffsig_eval.R
\name{diffsig_eval}
\alias{diffsig_eval}
\title{Diffsig coverage and root mean squared error (RMSE) calculation function}
\usage{
diffsig_eval(betatrue, fit, ci_level = 80, ...)
}
\arguments{
\item{betatrue}{a vector of true betas}

\item{fit}{a \code{stanfit} object returned by \code{rstan::stan}}

\item{ci_level}{a 2-dimension vector of lower/upper credible interval}

\item{...}{Arguments passed to \code{rstan::stan} (e.g. thin, init, ...).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::stan}
}
\description{
Diffsig coverage and root mean squared error (RMSE) calculation function
}
\examples{
data(simdat)
data(fit)
diffsig_eval(simdat$beta, fit, ci_level=80)

}
