% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffsig_plot.R
\name{diffsig_plot}
\alias{diffsig_plot}
\title{Plot posterior estimates results from diffsig_fit}
\usage{
diffsig_plot(fit, pars = "beta", rowlabels = NULL, ...)
}
\arguments{
\item{fit}{stan fit object from diffsig_stan}

\item{pars}{parameters to plot}

\item{rowlabels}{labels for betas in diffsig plot}

\item{...}{Arguments passed to ggplot}
}
\value{
Diffsig plot
}
\description{
Plot posterior estimates results from diffsig_fit
}
